+patient

+++ src/Controller/ReservationRdvController.php
@@ -2,16 +2,18 @@
 
 namespace App\Controller;
 
-use App\Entity\Medecin;
+use App\Entity\Med;
 use App\Entity\Patient;
 use App\Entity\ReservationRdv;
 use App\Form\ReservatioRdvType;
 use App\Form\SearchType;
 use App\Model\SearchMedecin;
-use App\Repository\MedecinRepository;
+use App\Repository\MedRepository;
+use App\Repository\PatientRepository;
 use App\Repository\ReservationRdvRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Knp\Component\Pager\PaginatorInterface;
+
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -28,7 +30,7 @@ class ReservationRdvController extends AbstractController
     //     ]);
     // }
     #[Route('/{id}/search', name: 'app_reservation_search')]
-    public function serachMedecin(Request $request, MedecinRepository $repMedecins, PaginatorInterface $paginatorInterface, int $id): Response
+    public function serachMedecin(Request $request, MedRepository $repMedecins, PaginatorInterface $paginatorInterface, int $id): Response
     {
         $serachMedecin = new SearchMedecin();
         $form = $this->createForm(SearchType::class, $serachMedecin);
@@ -54,7 +56,25 @@ class ReservationRdvController extends AbstractController
     public function listRdvPatient(ReservationRdvRepository $reservationRdvRepository, int $id): Response
     {
         $listesrdvs = $reservationRdvRepository->findBy(['patient' => $id]);
+        $rdvs = [];
+        foreach ($listesrdvs as $rdv) {
+            $rdvs[] = [
+
+                'id' => $rdv->getId(),
+                'start' => $rdv->getDateRdv()->format('Y-m-d H:i:s'),
+                'end' => $rdv->getDateRdv()->format('Y-m-d H:i:s'),
+                'title' => ' Rdv: ' . $rdv->getMotif(),
+                'description' => '',
+                'backgroundColor' => '#99c93d',
+                'borderColor' => '#99c93d',
+                'textColor' => 'white',
+                'allDay' => true,
+
+            ];
+        }
+        $data = json_encode($rdvs);
         return $this->render('reservation_rdv/listesPatient.html.twig', [
+            'data' => $data,
             'idPatient' => $id,
             'listesrdvs' => $listesrdvs,
         ]);
@@ -71,10 +91,10 @@ class ReservationRdvController extends AbstractController
     }
 
     #[Route('/{idMedecin}/{idPatient}/new', name: 'app_reservation_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager, int $idPatient, int $idMedecin): Response
+    public function new(Request $request, EntityManagerInterface $entityManager, int $idPatient, int $idMedecin,): Response
     {
         $patient = $entityManager->getRepository(Patient::class)->find($idPatient);
-        $medecin = $entityManager->getRepository(Medecin::class)->find($idMedecin);
+        $medecin = $entityManager->getRepository(Med::class)->find($idMedecin);
         $rdv = new ReservationRdv();
         $rdv->setPatient($patient);
         $rdv->setMedecin($medecin);
@@ -117,12 +137,17 @@ class ReservationRdvController extends AbstractController
         ]);
     }
 
-    #[Route('/delete/{idPatient}/{id}', name: 'app_reservation_delete', methods: ['POST'])]
-    public function delete(EntityManagerInterface $entityManager, ReservationRdvRepository $rep, $id, $idPatient): Response
+    #[Route('/delete/{idPatient}/{id}', name: 'app_reservation_delete', methods: ['GET', 'POST'])]
+    public function delete(EntityManagerInterface $entityManager, PatientRepository $rep1, ReservationRdvRepository $rep2, $id, $idPatient): Response
     {
-        $rdv = $rep->find($id);
-        $entityManager->remove($rdv);
-        $entityManager->flush();
+        $rdv = $rep2->find($id);
+        if ($rdv !== null) {
+            $p = $rep1->find($idPatient);
+            $nbrAnnulations = $p->getNbAnnulations();
+            $p->setNbAnnulations($nbrAnnulations + 1);
+            $entityManager->remove($rdv);
+            $entityManager->flush();
+        }
         return $this->redirectToRoute('app_reservation_listes_patient', ['id' => $idPatient,], Response::HTTP_SEE_OTHER);
     }
 
@@ -131,7 +156,7 @@ class ReservationRdvController extends AbstractController
     #[Route('/listRdv/{idMedecin}', name: 'app_reservation_listes_medecin')]
     public function rdvList(EntityManagerInterface $entityManager, ReservationRdvRepository $reservationRepository, $idMedecin): Response
     {
-        $medecin = $entityManager->getRepository(Medecin::class)->find($idMedecin);
+        $medecin = $entityManager->getRepository(Med::class)->find($idMedecin);
         $reservations = $reservationRepository->findBy(['medecin' => $medecin]);
 
         return $this->render('reservation_rdv/listesMedecin.html.twig', [
